requirements.txt
flask
opencv-python     
requests
Pillow
paho-mqtt           
mtcnn
tensorflow